requirements.txt
bleach==6.2.0
build==1.2.2.post1
Django==5.1.3
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
docutils==0.21.2
environ==1.0
folium==0.19.4
frozenlist==1.5.0
installer==0.7.0
instaloader==4.14.1
poetry==2.1.1
poetry-core==2.1.1
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pymunk==6.9.0
PyMuPDF==1.25.3
pynput==1.7.7
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytube==15.0.0
selenium==4.29.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
selenium-stealth==1.0.6
Pillow>=10.0
gunicorn==20.1.0  # Укажи актуальную версию gunicorn
whitenoise==6.3.0  # укажи актуальную версию
PyPDF2==3.0.1



# new
django-ckeditor==6.7.2
unidecode==1.3.8